requirements.txt
psycopg2-binary
fastapi
uvicorn
redis